Automatically adjust the name of the local repo in the pacman.conf of both containers to the name set in data/settings.ini.
This gets the var pacmanrepo_name from data/settings.ini. If none is
provided, abort the startup. The initial config has the default set.
This change is done in the startup script and not in Djano on purpose
to possibly allow running Django as non-root in the future.
Regeneration of the makepkg container has to be triggered manually after
changing the repo name. TODO: Find a good way to do this automatically.