cd_manager: Don't use a list as a default arg in the constructor.
Package.build() had a list as default arg in the constructor that isn't needed for
years but may cause weird bugs. A default list in the constructor is a
static pointer. This is a weird edge-case of CPython and
can be solved simply by instating the list in the function body.